SoundPackManagerDialog: move AlertCategoryItem + FRIENDLY_ALERT_CATEGORIES into soundpack_manager/constants and import. Shrink dialog file further. Tests passing.
